refactor: 비즈니스 로직 서비스 코드로 이동
