+++ src/main/java/com/nexters/goalpanzi/application/mission/MissionMemberService.java
@@ -1,8 +1,10 @@
 package com.nexters.goalpanzi.application.mission;
 
+import com.nexters.goalpanzi.application.firebase.TopicGenerator;
 import com.nexters.goalpanzi.application.mission.dto.response.MemberRankResponse;
 import com.nexters.goalpanzi.application.mission.dto.response.MissionDetailResponse;
 import com.nexters.goalpanzi.application.mission.dto.response.MissionsResponse;
+import com.nexters.goalpanzi.application.mission.event.JoinMissionEvent;
 import com.nexters.goalpanzi.domain.common.BaseEntity;
 import com.nexters.goalpanzi.domain.member.Member;
 import com.nexters.goalpanzi.domain.member.repository.MemberRepository;
@@ -12,12 +14,16 @@ import com.nexters.goalpanzi.domain.mission.repository.MissionRepository;
 import com.nexters.goalpanzi.exception.AlreadyExistsException;
 import com.nexters.goalpanzi.exception.ErrorCode;
 import com.nexters.goalpanzi.exception.NotFoundException;
+import com.nexters.goalpanzi.infrastructure.firebase.PushNotificationSender;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+
+import static com.nexters.goalpanzi.domain.firebase.PushNotificationMessage.MISSION_CANCELLATION_WARNING;
+import static com.nexters.goalpanzi.domain.firebase.PushNotificationMessage.MISSION_READY;
 
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
@@ -31,6 +37,7 @@ public class MissionMemberService {
     private final MemberRepository memberRepository;
 
     private final ApplicationEventPublisher eventPublisher;
+    private final PushNotificationSender pushNotificationSender;
 
     public MissionDetailResponse getJoinableMission(final InvitationCode invitationCode) {
         missionValidator.validateJoinableMission(invitationCode);
@@ -45,8 +52,9 @@ public class MissionMemberService {
         missionValidator.validateMaxPersonnel(mission);
         missionMemberRepository.save(MissionMember.join(member, mission));
 
-//        TODO
-//        eventPublisher.publishEvent(new JoinMissionEvent(mission.getId(), "TODO deviceToken", member.getNickname()));
+        if (member.getDeviceToken() != null) {
+            eventPublisher.publishEvent(new JoinMissionEvent(mission.getId(), member.getDeviceToken(), member.getNickname()));
+        }
     }
 
     private Mission getMissionByCode(final InvitationCode invitationCode) {
@@ -118,4 +126,42 @@ public class MissionMemberService {
         MissionMember missionMember = missionMemberRepository.getMissionMember(memberId, missionId);
         missionMember.checkCompleted();
     }
+
+    @Transactional
+    public void sendReadyPushMessage() {
+        List<Mission> missions = missionRepository.getReadyMissions();
+        missions.forEach(mission -> {
+            if (mission.isReadyTime() && hasEnoughMember(mission.getId())) {
+                String topic = TopicGenerator.getTopic(mission.getId());
+                pushNotificationSender.sendGroupMessage(
+                        MISSION_READY.getTitle(),
+                        MISSION_READY.getBody(),
+                        topic
+                );
+            }
+        });
+    }
+
+    @Transactional
+    public void sendCancellationWarningPushMessage() {
+        List<Mission> missions = missionRepository.getReadyMissions();
+        missions.forEach(mission -> {
+            if (mission.isReadyTime() && !hasEnoughMember(mission.getId())) {
+                String topic = TopicGenerator.getTopic(mission.getId());
+                pushNotificationSender.sendGroupMessage(
+                        MISSION_CANCELLATION_WARNING.getTitle(),
+                        MISSION_CANCELLATION_WARNING.getBody(),
+                        topic
+                );
+            }
+        });
+    }
+
+    private boolean hasEnoughMember(final Long missionId) {
+        List<MissionMember> missionMembers = missionMemberRepository.findAllByMissionId(missionId);
+        int memberCount = missionMembers.size();
+
+        return memberCount >= MissionMemberCount.MIN.getCount();
+    }
+
 }
